Pickup and droppen van Pickup verbeterd
-Code_GameManager, now communicates with the Code_PickUpManager. So it
can send it's arena to that of the Code_PickUpManager

-Code_Shield, the OnTriggerEnter no longer checks if the shield hits a
pickup Object

-Code_Player, changed the deathDel into a bool, So the Code_PickUp can
do easier checks

-Code_PickUpManager, Removed the old dropPoints system, now the pickups
drop in a more dynamic way and are less predictable with the dropRadius
system

-Code_PickUp,  Removed the repool system, the objects no longer repool
themselves, added a isUnpooled bool so the pickUpMng can keep track of
it, Added a CheckPlayerForPowerUp virtual method. Now the pickup object
will check if the player already has it's effect, so it won't stack or
will given the pickup. Added the OnTriggerEnter so the player can now
walk through the pickup to activate it. This was done by editing the
Editor physics by creating two more layers. One for the player and one
for pickups

-Code_PickUp_Shield, Overrides the CheckPlayerForPickUp method and asks
the player if he already has a deathshield

-Code_Arena, now has a GetCurrentRing method so other scripts can ask
what it's current ring count is, It also checks if a pickup object fell
off the arena and pools it if this is the case

-GameScene, Changed and added some new rocks. All rocks are now convexed
and I created ramps, so the players can get up higher places