CameraControl
The zoom is a bit annoying, but tomorow i'll try to fix it. Gotta use
percentages of the total at the start of the game and then compare the
100 percent with the updated diferences and zoom in based on the outcome
of the difference in percentages.

Or we could just make the camera move to a new location whenever a ring
falls.